Add files via upload
This script can help you convert audio to text. Its output is a txt file in the folder where the script is located.

You can pass the name of the audio file through the command line if it is in the same script directory, it is also possible to pass the full file path.